Major Fix to Main
point_spectra_gui/functions/MainWindow.py:1: Adding necessary imports
from original main
point_spectra_gui/functions/MainWindow.py:60: Moving self.setting to
Basics
point_spectra_gui/functions/MainWindow.py:81: Trying to add debug mode
memory to application
point_spectra_gui/functions/MainWindow.py:191: we can create new
applications from within MainWindow.py
point_spectra_gui/functions/MainWindow.py:360: Create the splash screen
as usual.
point_spectra_gui/functions/MainWindow.py:386 Because we are within
MainWindow.py, we will be now starting the application locally instead
of referencing like we were before
